+++ frontend/src/data/config.js
@@ -1,5 +1,32 @@
-import { Web3Storage } from 'web3.storage'
+import { Web3Storage } from 'web3.storage';
+import govData from "./ContractABIs/Governance.json";
+import nftData from "./ContractABIs/MembershipNFT.json";
+import tokenData from "./ContractABIs/UDAOToken.json";
 
 function web3AccessToken () {
     return process.env.WEB3STORAGE_TOKEN
+}
+
+export function govAddress() {
+    return '0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0';
+}
+
+export function nftAddress() {
+    return '0x5FbDB2315678afecb367f032d93F642f64180aa3';
+}
+
+export function tokenAddress() {
+    return '0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512';
+}
+
+export function govAbi() {
+    return govData.abi;
+}
+
+export function nftAbi() {
+    return nftData.abi;
+}
+
+export function tokenAbi() {
+    return tokenData.abi;
 }
